AdminCommand is now an interface

+++ src/command/commands/admin/DeleteCourseChannels.java
@@ -1,13 +1,14 @@
 package command.commands.admin;
 
 import command.AdminCommand;
+import command.Command;
 import net.dv8tion.jda.api.entities.Category;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 import java.util.Objects;
 
-public class DeleteCourseChannels extends AdminCommand {
+public class DeleteCourseChannels extends Command implements AdminCommand {
 
     /**
      * Initializes the command's key to "!deletecoursechannels"
@@ -17,7 +18,10 @@ public class DeleteCourseChannels extends AdminCommand {
     }
 
     @Override
-    protected void runCommand(MessageReceivedEvent event) {
+    public void start(MessageReceivedEvent event) {
+        if (!AdminCommand.memberIsAdmin(event.getMember()))
+            return;
+
         Category courses = Objects.requireNonNull(event.getGuild().getCategoriesByName("your channels", true).get(0));
         for (TextChannel channel : courses.getTextChannels())
             channel.delete().queue();
